testing/mochitest/tests/Harness_sanity/test_add_task.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for mochitest add_task, found in SpawnTask.js</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SpawnTask.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=">Mozilla Bug 1187701</a>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test">
<script class="testbody" type="text/javascript">

// Check that we can 'add_task' a few times and all tasks run asynchronously before test finishes.

add_task(async function() {
  var x = await Promise.resolve(1);
  is(x, 1, "task yields Promise value as expected");
});

add_task(async function() {
  var x = await [Promise.resolve(1), Promise.resolve(2), Promise.resolve(3)];
  is(x.join(""), "123", "task yields Promise value as expected");
});

add_task(async function() {
  var x = await (function* () {
    return 3;
  }());
  is(x, 3, "task yields generator function return value as expected");
});
</script>
</pre>
</body>
</html>
